pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
# CRITICAL: Changed name to match your new package name
name = "aicodeprep-gui" 
version = "1.2.3"
description = "A smart GUI to select, filter, and concatenate code files for large language model (LLM) and AI analysis."
readme = "README.md"
authors = [{ name = "Tom Brothers", email = "tom@wuu73.org" }]
license = { file = "SUSTAINABLE-LICENSE" }
classifiers = [
    # NOTE: You might want to change "Proprietary" to an open-source license like "MIT"
    # if you intend for others to use and contribute.
    "License :: Other/Proprietary License",  # Changed from OSI approved
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Software Development",
    "Environment :: X11 Applications :: Qt",
    "Environment :: Win32 (MS Windows)",
    "Environment :: MacOS X",
]
keywords = ["ai", "code", "preparation", "llm", "gui", "pyside6"]
dependencies = [
    # These are the RUNTIME dependencies. Users who pip install your package need these.
    "PySide6>=6.9,<6.10", # Simplified dependency, Essentials and Addons are included
    "toml",
    "NodeGraphQt",
    "pathspec",
    "requests", # for your network check
    "packaging",
    "Pygments>=2.18.0", # for syntax highlighting in preview window
    "litellm>=1.40.0", # for LLM provider integration in flow studio
    "setuptools>=65.0.0" # Required for Python 3.12+ compatibility (distutils was removed)
]
requires-python = ">=3.9,<3.14" # PySide6 6.9+ requires Python 3.9-3.13

[project.scripts]
# CRITICAL: Changed script name and entry point module
aicodeprep-gui = "aicodeprep_gui.main:main"
aicp = "aicodeprep_gui.main:main"

[project.urls]
# SUGGESTION: Update this once you rename the GitHub repo
Homepage = "https://github.com/detroittommy879/aicodeprep-gui" 

[tool.setuptools.packages.find]
# This is correct and will find your `aicodeprep_gui` package
where = ["."] 

[tool.setuptools.package-data]
# CRITICAL: Changed key to match your actual package name `aicodeprep_gui`
aicodeprep_gui = [
    "data/default_config.toml",
    "data/flow.json",
    "data/flow_studio_help.html",
    "images/*.png",
    "images/*.ico",
    "data/AICodePrep.workflow.zip",
    "data/fonts/*.ttf"
]

# SUGGESTION: Create optional dependencies for developers/builders
[project.optional-dependencies]
build = [
    "pyinstaller>=6.0.0",
    "pyinstaller-hooks-contrib",
    "setuptools",
    "wheel"
]
